requirements.txt
fastapi
uvicorn
python-dotenv
langchain-community
pydantic
faiss-cpu
gunicorn
starlette
langchain-google-genai
langchain-cohere
pypdf
google-generativeai